feat(curate): drop Test-an-idea section; unify tag sentiment on Taste Profile (#191, #189)
- Remove Curate 'Test an idea' (hypothesis): the CURATE_SECTIONS entry,
  CurateHypothesis and HypothesisVerdict components, their CSS, and the
  whole dead curation-batch backend (get_curation_batch,
  submit_curation_ratings, get_curation_verdict, get_tag_context_candidates)
  in Python and Go, plus their tests. The pair-stream tag filter stays.
- Tag sentiment is now one surface: Manage → Taste Profile is canonical
  (inferred + direct answers); Curate's section list links there instead of
  mounting a second copy of TagSentimentControl (#189).
