+++ core/backend.go
@@ -227,14 +227,6 @@ func dispatch(pluginDir string, payload jVal) (jVal, error) {
 		return opGetInspectorEntity(pluginDir, payload)
 	case "get_tag_sentiment_follow_up":
 		return opGetTagSentimentFollowUp(pluginDir, payload)
-	case "get_curation_batch":
-		return opGetCurationBatch(pluginDir, payload)
-	case "submit_curation_ratings":
-		return opSubmitCurationRatings(pluginDir, payload)
-	case "get_curation_verdict":
-		return opGetCurationVerdict(pluginDir, payload)
-	case "get_tag_context_candidates":
-		return opGetTagContextCandidates(pluginDir, payload)
 	case "get_curation_picks":
 		return opGetCurationPicks(pluginDir, payload)
 	case "submit_curation_picks":
