+++ core/curation.go
@@ -1,10 +1,8 @@
-// Curation ops: get_curation_batch, submit_curation_ratings,
-// get_curation_verdict, and get_tag_context_candidates — the Go mirror of
-// curator/curation.py. Hypothesis mode draws a stratified 2x2 sample around a
-// (base tag x context tag) pair with calibration anchors; explore mode
-// greedily maximizes NEW interactive-tag coverage (rarity-weighted, studio
-// penalty). Ratings live in feedback as feedback_type='curation_rating' with
-// the batch/cell in payload_json; verdicts use a batch's own ratings only.
+// Curation ops: pairwise picks, verdicts, and model impact — the Go mirror
+// of curator/curation.py. The pair loop surfaces deterministic pairs of
+// unlabeled scenes across the tag, performer, studio, and orthogonal
+// dimensions; picks write winner/loser labels into feedback that feed the
+// model, and pair verdicts accumulate per dimension across rounds.
 // Everything is deterministic (ORDER BY, sorted iteration, no RNG) so the
 // differential gates compare byte-identically against Python.
 package main
@@ -15,55 +13,16 @@ import (
 	"fmt"
 	"math"
 	"sort"
-	"strconv"
 	"strings"
 )
 
-const (
-	curationMinBudget                 = 1
-	curationMaxBudget                 = 40
-	curationDefaultBudget             = 20
-	curationRatingMin                 = 0
-	curationRatingMax                 = 10
-	curationConfirmDelta              = 0.15
-	curationConfirmMinN               = 10
-	curationAnchorBandSize            = 200
-	curationMaxItemTags               = 8
-	curationDefaultMinSupport         = 20
-	curationContrastMinLabeled        = 4
-	curationExploreAnchors            = 3
-	curationHypothesisAnchorFraction  = 0.15
-	curationHypothesisControlFraction = 0.15
-	curationMaxLibraryRate            = 0.30
-	curationContrastEvidenceScale     = 8.0
-)
-
-// suggestionExcludedCategories mirrors curation.SUGGESTION_EXCLUDED_CATEGORIES.
-var suggestionExcludedCategories = map[string]bool{
-	"Clothing": true, "Moods": true, "Locations": true, "Shot Type": true,
-	"Surfaces": true, "Misc": true, "Accessories": true,
-}
+const curationMaxItemTags = 8
 
 // curationExcludedCategories mirrors curation.EXCLUDED_CATEGORIES.
 var curationExcludedCategories = map[string]bool{
 	"Hair Color": true, "Hair Style": true, "Body Type": true, "Breasts": true,
 	"Face": true, "Skin Tone": true, "Piercings": true, "Ass": true,
 	"Genitals": true, "Height": true, "Tattoos": true, "Race": true,
-}
-
-// curationReasonTypes mirrors curation.REASON_TYPES.
-var curationReasonTypes = map[string]bool{
-	"metadata_wrong": true, "not_now": true, "contradicts_hypothesis": true,
-	"performer_driven": true,
-}
-
-// curationHypothesisCells mirrors curation.HYPOTHESIS_CELLS.
-var curationHypothesisCells = []string{"L&T", "L&!T", "!L&T", "!L&!T", "anchor"}
-
-type curationItem struct {
-	sceneID string
-	cell    string
-	anchor  bool
 }
 
 type curationContext struct {
@@ -96,21 +55,6 @@ func (c *curationContext) rarity(tagID string) float64 {
 
 func (c *curationContext) isInteractive(tagID string) bool {
 	return c.interactive[tagID]
-}
-
-func curationHalfUp(value float64) int64 {
-	return int64(value + 0.5)
-}
-
-func curationMean(values []float64) float64 {
-	if len(values) == 0 {
-		return 0.0
-	}
-	var total float64
-	for _, v := range values {
-		total += v
-	}
-	return total / float64(len(values))
 }
 
 // loadCurationContext mirrors curation.curation_context.
@@ -394,85 +338,11 @@ ORDER BY published_at_ms DESC LIMIT 1`).Scan(&modelID); err != nil && err != sql
 	return ctx, nil
 }
 
-func curationUnlabeledPool(ctx *curationContext, scenes map[string]bool) []string {
-	var out []string
-	for sceneID := range scenes {
-		if !ctx.labels[sceneID] && !ctx.blockedScenes[sceneID] && !ctx.metadataWrong[sceneID] {
-			out = append(out, sceneID)
-		}
-	}
-	sort.Strings(out)
-	return out
-}
-
 func curationStudioName(ctx *curationContext, sceneID string) string {
 	if name := ctx.studio[sceneID]; name != "" {
 		return name
 	}
 	return "?"
-}
-
-// curationRoundRobin mirrors curation._round_robin: one scene per studio
-// before repeats, deterministic order.
-func curationRoundRobin(pool []string, quota int, ctx *curationContext) []string {
-	byStudio := map[string][]string{}
-	for _, sceneID := range pool {
-		name := curationStudioName(ctx, sceneID)
-		byStudio[name] = append(byStudio[name], sceneID)
-	}
-	names := make([]string, 0, len(byStudio))
-	for name, sids := range byStudio {
-		sort.Strings(sids)
-		names = append(names, name)
-	}
-	sort.Strings(names)
-	var chosen []string
-	for _, name := range names {
-		if len(chosen) >= quota {
-			break
-		}
-		chosen = append(chosen, byStudio[name][0])
-		byStudio[name] = byStudio[name][1:]
-	}
-	for len(chosen) < quota {
-		progressed := false
-		for _, name := range names {
-			if len(chosen) >= quota {
-				break
-			}
-			if len(byStudio[name]) > 0 {
-				chosen = append(chosen, byStudio[name][0])
-				byStudio[name] = byStudio[name][1:]
-				progressed = true
-			}
-		}
-		if !progressed {
-			break
-		}
-	}
-	return chosen
-}
-
-// curationAnchorBand mirrors curation._anchor_band: the appeal middle band.
-func curationAnchorBand(pool []string, ctx *curationContext) []string {
-	ordered := append([]string(nil), pool...)
-	sort.Slice(ordered, func(i, j int) bool {
-		a, b := ctx.appeal[ordered[i]], ctx.appeal[ordered[j]]
-		if a != b {
-			return a < b
-		}
-		return ordered[i] < ordered[j]
-	})
-	mid := len(ordered) / 2
-	lo := mid - curationAnchorBandSize
-	if lo < 0 {
-		lo = 0
-	}
-	hi := mid + curationAnchorBandSize
-	if hi > len(ordered) {
-		hi = len(ordered)
-	}
-	return ordered[lo:hi]
 }
 
 func curationItemTags(ctx *curationContext, sceneID string) jVal {
@@ -517,1043 +387,9 @@ func curationItemTags(ctx *curationContext, sceneID string) jVal {
 	return out
 }
 
-// curationSelectHypothesis mirrors curation.select_hypothesis.
-func curationSelectHypothesis(
-	ctx *curationContext, baseTag, contextTag string, budget int,
-) ([]curationItem, map[string]int64) {
-	base := map[string]bool{}
-	context := map[string]bool{}
-	for sceneID, tags := range ctx.sceneTags {
-		if tags[baseTag] {
-			base[sceneID] = true
-		}
-		if tags[contextTag] {
-			context[sceneID] = true
-		}
-	}
-	pools := map[string][]string{}
-	var poolKeys []string
-	for _, cell := range []struct {
-		name   string
-		scenes map[string]bool
-	}{
-		{"L&T", intersect(base, context)},
-		{"L&!T", difference(base, context)},
-		{"!L&T", difference(context, base)},
-		{"!L&!T", difference(ctx.sceneIDs, union(base, context))},
-	} {
-		pool := curationUnlabeledPool(ctx, cell.scenes)
-		if cell.name == "L&!T" {
-			pool = groupCleanScenes(ctx, pool, contextTag)
-		}
-		pools[cell.name] = pool
-		poolKeys = append(poolKeys, cell.name)
-	}
-	anchors := curationHalfUp(float64(budget) * curationHypothesisAnchorFraction)
-	if anchors < 1 {
-		anchors = 1
-	}
-	controls := curationHalfUp(float64(budget) * curationHypothesisControlFraction)
-	if controls < 1 {
-		controls = 1
-	}
-	if controls > int64(budget)-anchors {
-		controls = int64(budget) - anchors
-	}
-	contrast := int64(budget) - anchors - controls
-	contrastLT := contrast/2 + contrast%2
-	contrastLNT := contrast / 2
-	items := []curationItem{}
-	for _, cell := range []struct {
-		name  string
-		quota int64
-	}{
-		{"L&T", contrastLT},
-		{"L&!T", contrastLNT},
-		{"!L&T", controls},
-	} {
-		for _, sceneID := range curationRoundRobin(pools[cell.name], int(cell.quota), ctx) {
-			items = append(items, curationItem{sceneID: sceneID, cell: cell.name})
-		}
-	}
-	anchorPool := curationAnchorBand(pools["!L&!T"], ctx)
-	for _, sceneID := range curationRoundRobin(anchorPool, int(anchors), ctx) {
-		items = append(items, curationItem{sceneID: sceneID, cell: "anchor", anchor: true})
-	}
-	sort.Slice(items, func(i, j int) bool {
-		if items[i].cell != items[j].cell {
-			return items[i].cell < items[j].cell
-		}
-		return items[i].sceneID < items[j].sceneID
-	})
-	pool := map[string]int64{}
-	for _, name := range poolKeys {
-		pool[name] = int64(len(pools[name]))
-	}
-	return items, pool
-}
-
-func intersect(a, b map[string]bool) map[string]bool {
-	out := map[string]bool{}
-	for k := range a {
-		if b[k] {
-			out[k] = true
-		}
-	}
-	return out
-}
-
-func difference(a, b map[string]bool) map[string]bool {
-	out := map[string]bool{}
-	for k := range a {
-		if !b[k] {
-			out[k] = true
-		}
-	}
-	return out
-}
-
-func union(a, b map[string]bool) map[string]bool {
-	out := map[string]bool{}
-	for k := range a {
-		out[k] = true
-	}
-	for k := range b {
-		out[k] = true
-	}
-	return out
-}
-
-// curationSelectExplore mirrors curation.select_explore.
-func curationSelectExplore(
-	ctx *curationContext, budget int,
-) ([]curationItem, map[string]int64) {
-	pool := curationUnlabeledPool(ctx, ctx.sceneIDs)
-	anchors := curationExploreAnchors
-	if budget/2 < anchors {
-		anchors = budget / 2
-	}
-	exploreBudget := budget - anchors
-	tagScenes := map[string][]string{}
-	for _, sceneID := range pool {
-		for tagID := range ctx.sceneTags[sceneID] {
-			if ctx.isInteractive(tagID) {
-				tagScenes[tagID] = append(tagScenes[tagID], sceneID)
-			}
-		}
-	}
-	value := map[string]float64{}
-	for _, sceneID := range pool {
-		var total float64
-		for tagID := range ctx.sceneTags[sceneID] {
-			if ctx.isInteractive(tagID) {
-				total += ctx.rarity(tagID)
-			}
-		}
-		value[sceneID] = total
-	}
-	covered := map[string]bool{}
-	chosenSet := map[string]bool{}
-	chosenStudios := map[string]bool{}
-	var chosen []string
-	for len(chosen) < exploreBudget {
-		best, bestValue := "", -1.0
-		for _, sceneID := range pool {
-			if chosenSet[sceneID] {
-				continue
-			}
-			v := value[sceneID]
-			if chosenStudios[curationStudioName(ctx, sceneID)] {
-				v *= 0.5
-			}
-			if v > bestValue {
-				best, bestValue = sceneID, v
-			}
-		}
-		if best == "" || bestValue <= 0 {
-			break
-		}
-		chosen = append(chosen, best)
-		chosenSet[best] = true
-		chosenStudios[curationStudioName(ctx, best)] = true
-		for tagID := range ctx.sceneTags[best] {
-			if ctx.isInteractive(tagID) && !covered[tagID] {
-				covered[tagID] = true
-				for _, other := range tagScenes[tagID] {
-					if _, ok := value[other]; ok {
-						value[other] -= ctx.rarity(tagID)
-					}
-				}
-			}
-		}
-	}
-	var anchorPool []string
-	for _, sceneID := range pool {
-		if !chosenSet[sceneID] {
-			anchorPool = append(anchorPool, sceneID)
-		}
-	}
-	items := []curationItem{}
-	for _, sceneID := range chosen {
-		items = append(items, curationItem{sceneID: sceneID, cell: "explore"})
-	}
-	for _, sceneID := range curationRoundRobin(curationAnchorBand(anchorPool, ctx), anchors, ctx) {
-		items = append(items, curationItem{sceneID: sceneID, cell: "anchor", anchor: true})
-	}
-	sort.Slice(items, func(i, j int) bool {
-		if items[i].cell != items[j].cell {
-			return items[i].cell < items[j].cell
-		}
-		return items[i].sceneID < items[j].sceneID
-	})
-	return items, map[string]int64{
-		"candidates":       int64(len(pool)),
-		"interactive_tags": int64(len(ctx.interactive)),
-	}
-}
-
-// createCurationBatch mirrors curation.create_batch.
-func createCurationBatch(db dbx, mode, baseTagID, contextTagID string, budget int) (jVal, error) {
-	if mode != "hypothesis" && mode != "explore" {
-		return jvNull(), fmt.Errorf("mode must be 'hypothesis' or 'explore'")
-	}
-	if budget < curationMinBudget || budget > curationMaxBudget {
-		return jvNull(), fmt.Errorf("budget must be from %d to %d", curationMinBudget, curationMaxBudget)
-	}
-	if mode == "hypothesis" {
-		if baseTagID == "" || contextTagID == "" {
-			return jvNull(), fmt.Errorf("hypothesis mode requires base_tag_id and context_tag_id")
-		}
-		if baseTagID == contextTagID {
-			return jvNull(), fmt.Errorf("base_tag_id and context_tag_id must differ")
-		}
-	}
-	for _, tagID := range []string{baseTagID, contextTagID} {
-		if tagID == "" {
-			continue
-		}
-		var probe int
-		err := db.QueryRow(`SELECT 1 FROM source_tag WHERE tag_id=?`, tagID).Scan(&probe)
-		if err == sql.ErrNoRows {
-			return jvNull(), fmt.Errorf("unknown tag: %s", tagID)
-		}
-		if err != nil {
-			return jvNull(), err
-		}
-	}
-	ctx, err := loadCurationContext(db)
-	if err != nil {
-		return jvNull(), err
-	}
-	var items []curationItem
-	var pool map[string]int64
-	var policy string
-	if mode == "hypothesis" {
-		items, pool = curationSelectHypothesis(ctx, baseTagID, contextTagID, budget)
-		anchors := 0
-		for _, item := range items {
-			if item.anchor {
-				anchors++
-			}
-		}
-		policy = fmt.Sprintf(
-			"stratified 2x2, studio round-robin, unlabeled only, %d calibration anchors",
-			anchors,
-		)
-	} else {
-		items, pool = curationSelectExplore(ctx, budget)
-		policy = "max-coverage interactive tags, rarity-weighted, studio penalty, unlabeled only"
-	}
-	batchID := uuid4()
-	now := nowMs()
-	payloadJSON, err := marshalSorted(policy, pool)
-	if err != nil {
-		return jvNull(), err
-	}
-	err = withTxn(db, func(conn *sql.Conn) error {
-		var baseArg, contextArg any
-		if baseTagID != "" {
-			baseArg = baseTagID
-		}
-		if contextTagID != "" {
-			contextArg = contextTagID
-		}
-		if _, err := conn.ExecContext(context.Background(), `
-INSERT INTO curation_batch(
-    batch_id, mode, base_tag_id, context_tag_id, budget, status,
-    created_at_ms, payload_json
-) VALUES (?, ?, ?, ?, ?, 'open', ?, ?)`,
-			batchID, mode, baseArg, contextArg, budget, now, payloadJSON); err != nil {
-			return err
-		}
-		for _, item := range items {
-			anchor := 0
-			if item.anchor {
-				anchor = 1
-			}
-			if _, err := conn.ExecContext(context.Background(), `
-INSERT INTO curation_batch_item(batch_id, scene_id, cell, anchor)
-VALUES (?, ?, ?, ?)`, batchID, item.sceneID, item.cell, anchor); err != nil {
-				return err
-			}
-		}
-		return nil
-	})
-	if err != nil {
-		return jvNull(), err
-	}
-	itemsOut := jvArr()
-	for _, item := range items {
-		titleVal, studioVal, dateVal, detailsVal := jvNull(), jvNull(), jvNull(), jvNull()
-		if title := ctx.sceneTitle[item.sceneID]; title != "" {
-			titleVal = jvStr(title)
-		}
-		if studio := ctx.studio[item.sceneID]; studio != "" {
-			studioVal = jvStr(studio)
-		}
-		if date := ctx.sceneDate[item.sceneID]; date != "" {
-			dateVal = jvStr(date)
-		}
-		if details := ctx.sceneDetails[item.sceneID]; details != "" {
-			detailsVal = jvStr(details)
-		}
-		itemsOut.arr = append(itemsOut.arr, jvObj(
-			jvKey("scene_id", jvStr(item.sceneID)),
-			jvKey("cell", jvStr(item.cell)),
-			jvKey("anchor", jvBool(item.anchor)),
-			jvKey("title", titleVal),
-			jvKey("studio", studioVal),
-			jvKey("date", dateVal),
-			jvKey("details", detailsVal),
-			jvKey("tags", curationItemTags(ctx, item.sceneID)),
-		))
-	}
-	baseVal, contextVal := jvNull(), jvNull()
-	if baseTagID != "" {
-		baseVal = jvStr(baseTagID)
-	}
-	if contextTagID != "" {
-		contextVal = jvStr(contextTagID)
-	}
-	return jvObj(
-		jvKey("schema_version", jvInt(2)),
-		jvKey("batch_id", jvStr(batchID)),
-		jvKey("mode", jvStr(mode)),
-		jvKey("base_tag_id", baseVal),
-		jvKey("context_tag_id", contextVal),
-		jvKey("budget", jvInt(int64(budget))),
-		jvKey("items", itemsOut),
-		jvKey("pool", poolToJVal(pool)),
-		jvKey("policy", jvStr(policy)),
-	), nil
-}
-
-func marshalSorted(policy string, pool map[string]int64) (string, error) {
-	payload := jvObj(
-		jvKey("policy", jvStr(policy)),
-		jvKey("pool", poolToJVal(pool)),
-	)
-	return payload.marshalSortedKeys(), nil
-}
-
-func poolToJVal(pool map[string]int64) jVal {
-	out := jvObj()
-	for _, key := range sortedMapKeys(pool) {
-		out.obj = append(out.obj, jPair{key: key, val: jvInt(pool[key])})
-	}
-	return out
-}
-
-func curationBatchItems(db dbx, batchID string) (map[string]struct {
-	cell  string
-	rated bool
-}, error) {
-	rows, err := db.Query(
-		`SELECT scene_id, cell, rated FROM curation_batch_item WHERE batch_id=?`,
-		batchID,
-	)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-	out := map[string]struct {
-		cell  string
-		rated bool
-	}{}
-	for rows.Next() {
-		var sceneID, cell string
-		var rated int64
-		if err := rows.Scan(&sceneID, &cell, &rated); err != nil {
-			return nil, err
-		}
-		out[sceneID] = struct {
-			cell  string
-			rated bool
-		}{cell: cell, rated: rated != 0}
-	}
-	return out, rows.Err()
-}
-
-// submitCurationRatings mirrors curation.submit_ratings.
-func submitCurationRatings(db dbx, batchID string, ratings jVal) (jVal, error) {
-	if batchID == "" {
-		return jvNull(), fmt.Errorf("batch_id is required")
-	}
-	var status string
-	err := db.QueryRow(
-		`SELECT status FROM curation_batch WHERE batch_id=?`, batchID,
-	).Scan(&status)
-	if err == sql.ErrNoRows {
-		return jvNull(), fmt.Errorf("unknown batch: %s", batchID)
-	}
-	if err != nil {
-		return jvNull(), err
-	}
-	if status != "open" {
-		return jvNull(), fmt.Errorf("batch is not open: %s", status)
-	}
-	items, err := curationBatchItems(db, batchID)
-	if err != nil {
-		return jvNull(), err
-	}
-	if len(items) == 0 {
-		return jvNull(), fmt.Errorf("unknown batch: %s", batchID)
-	}
-	seen := map[string]bool{}
-	type normalizedRating struct {
-		sceneID   string
-		value     int64
-		reason    string
-		hasReason bool
-	}
-	var normalized []normalizedRating
-	if ratings.kind != jArr {
-		return jvNull(), fmt.Errorf("ratings must be a list")
-	}
-	for _, entry := range ratings.arr {
-		if entry.kind != jObj {
-			return jvNull(), fmt.Errorf("each rating must be an object with scene_id, value, and reason")
-		}
-		sceneID := pythonStrOrEmpty(entry.get("scene_id"))
-		if sceneID == "" {
-			return jvNull(), fmt.Errorf("scene is not in this batch: %s", sceneID)
-		}
-		item, ok := items[sceneID]
-		if !ok {
-			return jvNull(), fmt.Errorf("scene is not in this batch: %s", sceneID)
-		}
-		if seen[sceneID] {
-			return jvNull(), fmt.Errorf("duplicate rating for scene: %s", sceneID)
-		}
-		if item.rated {
-			return jvNull(), fmt.Errorf("scene already rated in this batch: %s", sceneID)
-		}
-		rawValue := entry.get("value")
-		if rawValue.kind == jBool || rawValue.kind != jNum || rawValue.kindName() != "int" {
-			return jvNull(), fmt.Errorf("rating value must be an integer")
-		}
-		value, err := strconv.ParseInt(rawValue.num, 10, 64)
-		if err != nil {
-			return jvNull(), fmt.Errorf("rating value must be an integer")
-		}
-		if value < curationRatingMin || value > curationRatingMax {
-			return jvNull(), fmt.Errorf("rating value must be from %d to %d",
-				curationRatingMin, curationRatingMax)
-		}
-		reason := entry.get("reason")
-		reasonStr := ""
-		hasReason := false
-		if reason.kind != jNull {
-			if reason.kind != jStr {
-				return jvNull(), fmt.Errorf("unknown rating reason: %s", curationValueStr(reason))
-			}
-			if !curationReasonTypes[reason.s] {
-				return jvNull(), fmt.Errorf("unknown rating reason: %s", reason.s)
-			}
-			reasonStr = reason.s
-			hasReason = true
-		}
-		seen[sceneID] = true
-		normalized = append(normalized, normalizedRating{
-			sceneID: sceneID, value: value, reason: reasonStr, hasReason: hasReason,
-		})
-	}
-	if len(normalized) == 0 {
-		return jvNull(), fmt.Errorf("ratings must not be empty")
-	}
-	now := nowMs()
-	err = withTxn(db, func(conn *sql.Conn) error {
-		for _, rating := range normalized {
-			item := items[rating.sceneID]
-			payload := jvObj(
-				jvKey("batch_id", jvStr(batchID)),
-				jvKey("cell", jvStr(item.cell)),
-			)
-			if _, err := conn.ExecContext(context.Background(), `
-INSERT INTO feedback(
-    feedback_id, scene_id, feedback_type, value, occurred_at_ms,
-    reversed_by_id, impression_id, payload_json
-) VALUES (?, ?, 'curation_rating', ?, ?, NULL, NULL, ?)`,
-				nowStrID(now), rating.sceneID, strconv.FormatInt(rating.value, 10), now,
-				payload.marshalSortedKeys()); err != nil {
-				return err
-			}
-			if rating.hasReason {
-				if _, err := conn.ExecContext(context.Background(), `
-INSERT INTO feedback(
-    feedback_id, scene_id, feedback_type, value, occurred_at_ms,
-    reversed_by_id, impression_id, payload_json
-) VALUES (?, ?, ?, NULL, ?, NULL, NULL, '{}')`,
-					nowStrID(now), rating.sceneID, rating.reason, now); err != nil {
-					return err
-				}
-			}
-			if _, err := conn.ExecContext(context.Background(), `
-UPDATE curation_batch_item SET rated=1 WHERE batch_id=? AND scene_id=?`,
-				batchID, rating.sceneID); err != nil {
-				return err
-			}
-		}
-		var remaining int64
-		if err := conn.QueryRowContext(context.Background(), `
-SELECT count(*) FROM curation_batch_item WHERE batch_id=? AND rated=0`,
-			batchID).Scan(&remaining); err != nil {
-			return err
-		}
-		if remaining == 0 {
-			if _, err := conn.ExecContext(context.Background(), `
-UPDATE curation_batch SET status='rated' WHERE batch_id=?`, batchID); err != nil {
-				return err
-			}
-		}
-		return nil
-	})
-	if err != nil {
-		return jvNull(), err
-	}
-	statusOut := "open"
-	remaining, err := curationRemaining(db, batchID)
-	if err != nil {
-		return jvNull(), err
-	}
-	if remaining == 0 {
-		statusOut = "rated"
-	}
-	return jvObj(
-		jvKey("schema_version", jvInt(2)),
-		jvKey("accepted", jvInt(int64(len(normalized)))),
-		jvKey("batch_status", jvStr(statusOut)),
-	), nil
-}
-
-func curationRemaining(db dbx, batchID string) (int64, error) {
-	var remaining int64
-	err := db.QueryRow(
-		`SELECT count(*) FROM curation_batch_item WHERE batch_id=? AND rated=0`,
-		batchID,
-	).Scan(&remaining)
-	return remaining, err
-}
-
 func nowStrID(now int64) string {
 	return fmt.Sprintf("%d-%s", now, uuid4())
 }
-
-// curationBatchRatings mirrors curation._batch_ratings. Ratings marked with a
-// reason (metadata_wrong or contradicts_hypothesis) are excluded: those scenes
-// are not valid instances of the cell they were sorted into.
-func curationBatchRatings(db dbx, batchID string) (map[string]float64, error) {
-	excluded := map[string]bool{}
-	rows, err := db.Query(`
-SELECT DISTINCT scene_id FROM feedback
-WHERE feedback_type IN ('metadata_wrong', 'contradicts_hypothesis', 'performer_driven')
-  AND reversed_by_id IS NULL`)
-	if err != nil {
-		return nil, err
-	}
-	for rows.Next() {
-		var sceneID string
-		if err := rows.Scan(&sceneID); err != nil {
-			rows.Close()
-			return nil, err
-		}
-		excluded[sceneID] = true
-	}
-	rows.Close()
-	if err := rows.Err(); err != nil {
-		return nil, err
-	}
-	out := map[string]float64{}
-	rows, err = db.Query(`
-SELECT scene_id, value, payload_json FROM feedback
-WHERE feedback_type='curation_rating' AND reversed_by_id IS NULL`)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var sceneID string
-		var value, payloadJSON sql.NullString
-		if err := rows.Scan(&sceneID, &value, &payloadJSON); err != nil {
-			return nil, err
-		}
-		if !value.Valid {
-			continue
-		}
-		if excluded[sceneID] {
-			continue
-		}
-		payload, err := parseJSON([]byte(payloadJSON.String))
-		if err != nil {
-			continue
-		}
-		if pythonStrOrEmpty(payload.get("batch_id")) != batchID {
-			continue
-		}
-		rating, err := strconv.ParseInt(value.String, 10, 64)
-		if err != nil || rating < curationRatingMin || rating > curationRatingMax {
-			continue
-		}
-		out[sceneID] = (float64(rating) - 5) / 5
-	}
-	return out, rows.Err()
-}
-
-// curationVerdict mirrors curation.verdict.
-func curationVerdict(db dbx, batchID string) (jVal, error) {
-	if batchID == "" {
-		return jvNull(), fmt.Errorf("batch_id is required")
-	}
-	var mode string
-	var baseTagID, contextTagID sql.NullString
-	err := db.QueryRow(`
-SELECT mode, base_tag_id, context_tag_id FROM curation_batch WHERE batch_id=?`,
-		batchID).Scan(&mode, &baseTagID, &contextTagID)
-	if err == sql.ErrNoRows {
-		return jvNull(), fmt.Errorf("unknown batch: %s", batchID)
-	}
-	if err != nil {
-		return jvNull(), err
-	}
-	items, err := curationBatchItems(db, batchID)
-	if err != nil {
-		return jvNull(), err
-	}
-	outcomes, err := curationBatchRatings(db, batchID)
-	if err != nil {
-		return jvNull(), err
-	}
-	if mode == "hypothesis" {
-		byCell := map[string]*struct {
-			values []float64
-		}{}
-		for _, cell := range curationHypothesisCells {
-			byCell[cell] = &struct{ values []float64 }{}
-		}
-		for sceneID, item := range items {
-			if outcome, ok := outcomes[sceneID]; ok {
-				byCell[item.cell].values = append(byCell[item.cell].values, outcome)
-			}
-		}
-		cells := jvArr()
-		for _, cell := range curationHypothesisCells {
-			values := byCell[cell].values
-			mean := jvNull()
-			if len(values) > 0 {
-				mean = jvFloat(curationMean(values))
-			}
-			cells.arr = append(cells.arr, jvObj(
-				jvKey("cell", jvStr(cell)),
-				jvKey("n", jvInt(int64(len(values)))),
-				jvKey("mean_outcome", mean),
-			))
-		}
-		contrast := jvObj()
-		if len(byCell["L&T"].values) > 0 && len(byCell["L&!T"].values) > 0 {
-			delta := curationMean(byCell["L&T"].values) - curationMean(byCell["L&!T"].values)
-			nTotal := int64(len(byCell["L&T"].values) + len(byCell["L&!T"].values))
-			contrast = jvObj(
-				jvKey("delta", jvFloat(delta)),
-				jvKey("n_total", jvInt(nTotal)),
-				jvKey("confirmed", jvBool(mathAbs(delta) >= curationConfirmDelta && nTotal >= curationConfirmMinN)),
-			)
-		}
-		suggested := jvNull()
-		if len(byCell["L&T"].values) > 0 {
-			mean := curationMean(byCell["L&T"].values)
-			value := float64(curationHalfUp(mean*2)) / 2.0
-			if value > 1.0 {
-				value = 1.0
-			}
-			if value < -1.0 {
-				value = -1.0
-			}
-			baseVal, contextVal := "", ""
-			if baseTagID.Valid {
-				baseVal = baseTagID.String
-			}
-			if contextTagID.Valid {
-				contextVal = contextTagID.String
-			}
-			suggested = jvObj(
-				jvKey("base_tag_id", jvStr(baseVal)),
-				jvKey("context_tag_id", jvStr(contextVal)),
-				jvKey("value", jvFloat(value)),
-			)
-		}
-		return jvObj(
-			jvKey("schema_version", jvInt(2)),
-			jvKey("batch_id", jvStr(batchID)),
-			jvKey("mode", jvStr(mode)),
-			jvKey("cells", cells),
-			jvKey("contrast", contrast),
-			jvKey("suggested_rule", suggested),
-		), nil
-	}
-	// explore mode
-	tagRows := map[string][]float64{}
-	ratedScenes := make([]string, 0, len(outcomes))
-	for sceneID := range outcomes {
-		ratedScenes = append(ratedScenes, sceneID)
-	}
-	sort.Strings(ratedScenes)
-	if len(ratedScenes) > 0 {
-		placeholders := strings.TrimSuffix(strings.Repeat("?,", len(ratedScenes)), ",")
-		args := make([]any, len(ratedScenes))
-		for i, sceneID := range ratedScenes {
-			args[i] = sceneID
-		}
-		rows, err := db.Query(
-			`SELECT scene_id, tag_id FROM scene_tag WHERE scene_id IN (`+placeholders+`) ORDER BY scene_id, tag_id`,
-			args...,
-		)
-		if err != nil {
-			return jvNull(), err
-		}
-		for rows.Next() {
-			var sceneID, tagID string
-			if err := rows.Scan(&sceneID, &tagID); err != nil {
-				rows.Close()
-				return jvNull(), err
-			}
-			tagRows[tagID] = append(tagRows[tagID], outcomes[sceneID])
-		}
-		rows.Close()
-		if err := rows.Err(); err != nil {
-			return jvNull(), err
-		}
-	}
-	ctx, err := loadCurationContext(db)
-	if err != nil {
-		return jvNull(), err
-	}
-	type tagEntry struct {
-		tagID string
-		mean  float64
-	}
-	var entries []tagEntry
-	for tagID, values := range tagRows {
-		if len(values) >= 2 {
-			entries = append(entries, tagEntry{tagID: tagID, mean: curationMean(values)})
-		}
-	}
-	sort.Slice(entries, func(i, j int) bool {
-		if entries[i].mean != entries[j].mean {
-			return entries[i].mean > entries[j].mean
-		}
-		return entries[i].tagID < entries[j].tagID
-	})
-	entryJVal := func(entry tagEntry) jVal {
-		category := jvNull()
-		if cat := ctx.tagCat[entry.tagID]; cat != "" {
-			category = jvStr(cat)
-		}
-		name := ctx.tagName[entry.tagID]
-		if name == "" {
-			name = entry.tagID
-		}
-		return jvObj(
-			jvKey("tag_id", jvStr(entry.tagID)),
-			jvKey("name", jvStr(name)),
-			jvKey("category", category),
-			jvKey("n", jvInt(int64(len(tagRows[entry.tagID])))),
-			jvKey("mean_outcome", jvFloat(entry.mean)),
-		)
-	}
-	top := jvArr()
-	for _, entry := range entries[:minInt(len(entries), 10)] {
-		top.arr = append(top.arr, entryJVal(entry))
-	}
-	bottom := jvArr()
-	for i := len(entries) - 1; i >= 0 && len(bottom.arr) < 10; i-- {
-		bottom.arr = append(bottom.arr, entryJVal(entries[i]))
-	}
-	values := make([]float64, 0, len(outcomes))
-	for _, outcome := range outcomes {
-		values = append(values, outcome)
-	}
-	summaryMean := jvNull()
-	if len(values) > 0 {
-		summaryMean = jvFloat(curationMean(values))
-	}
-	return jvObj(
-		jvKey("schema_version", jvInt(2)),
-		jvKey("batch_id", jvStr(batchID)),
-		jvKey("mode", jvStr(mode)),
-		jvKey("summary", jvObj(
-			jvKey("n", jvInt(int64(len(values)))),
-			jvKey("mean_outcome", summaryMean),
-		)),
-		jvKey("top_tags", top),
-		jvKey("bottom_tags", bottom),
-	), nil
-}
-
-// tagContextCandidatesBody mirrors curation.tag_context_candidates.
-func tagContextCandidatesBody(db dbx, tagID string, minSupport int) (jVal, error) {
-	if tagID == "" {
-		return jvNull(), fmt.Errorf("tag_id is required")
-	}
-	var probe int
-	err := db.QueryRow(`SELECT 1 FROM source_tag WHERE tag_id=?`, tagID).Scan(&probe)
-	if err == sql.ErrNoRows {
-		return jvNull(), fmt.Errorf("unknown tag: %s", tagID)
-	}
-	if err != nil {
-		return jvNull(), err
-	}
-	if minSupport < 1 {
-		return jvNull(), fmt.Errorf("min_support must be at least 1")
-	}
-	ctx, err := loadCurationContext(db)
-	if err != nil {
-		return jvNull(), err
-	}
-	labels, err := modelSceneLabels(db)
-	if err != nil {
-		return jvNull(), err
-	}
-	baseScenes := map[string]bool{}
-	for sceneID, tags := range ctx.sceneTags {
-		if tags[tagID] {
-			baseScenes[sceneID] = true
-		}
-	}
-	labeledBase := map[string]float64{}
-	for sceneID := range baseScenes {
-		if label, ok := labels[sceneID]; ok {
-			labeledBase[sceneID] = label.outcome
-		}
-	}
-	cooc := map[string]int64{}
-	for sceneID := range baseScenes {
-		for t := range ctx.sceneTags[sceneID] {
-			if t != tagID && ctx.isInteractive(t) {
-				cooc[t]++
-			}
-		}
-	}
-	rowsOut := jvArr()
-	type candidate struct {
-		tagID    string
-		cooccur  int64
-		contrast *float64
-	}
-	var candidates []candidate
-	for t, n := range cooc {
-		if n < int64(minSupport) {
-			continue
-		}
-		totalScenes := float64(len(ctx.sceneIDs))
-		if totalScenes < 1 {
-			totalScenes = 1
-		}
-		if float64(ctx.counts[t])/totalScenes >= curationMaxLibraryRate {
-			continue // ubiquitous tags cannot discriminate any relationship
-		}
-		if strings.HasPrefix(ctx.tagName[t], "[") {
-			continue // sync-artifact tags ([Timestamp: Synced]...) are junk
-		}
-		if suggestionExcludedCategories[ctx.tagCat[t]] {
-			continue // weak-interaction categories are not hypotheses
-		}
-		var withT, withoutT []float64
-		for sceneID, outcome := range labeledBase {
-			if ctx.sceneTags[sceneID][t] {
-				withT = append(withT, outcome)
-			} else if len(ctx.scenePerformers[sceneID]) < 3 || ctx.tagCat[t] != "Group Makeup" {
-				withoutT = append(withoutT, outcome)
-			}
-		}
-		cand := candidate{tagID: t, cooccur: n}
-		if len(labeledBase) >= curationContrastMinLabeled && len(withT) > 0 && len(withoutT) > 0 {
-			raw := curationMean(withT) - curationMean(withoutT)
-			evidence := float64(minInt(len(withT), len(withoutT)))
-			factor := evidence / curationContrastEvidenceScale
-			if factor > 1.0 {
-				factor = 1.0
-			}
-			delta := raw * factor
-			cand.contrast = &delta
-		}
-		candidates = append(candidates, cand)
-	}
-	sort.Slice(candidates, func(i, j int) bool {
-		a, b := candidates[i], candidates[j]
-		aPos := a.contrast != nil && *a.contrast > 0
-		bPos := b.contrast != nil && *b.contrast > 0
-		if aPos != bPos {
-			return aPos
-		}
-		aC, bC := 0.0, 0.0
-		if a.contrast != nil {
-			aC = *a.contrast
-		}
-		if b.contrast != nil {
-			bC = *b.contrast
-		}
-		if aC != bC {
-			return aC > bC
-		}
-		if a.cooccur != b.cooccur {
-			return a.cooccur > b.cooccur
-		}
-		return ctx.tagName[a.tagID] < ctx.tagName[b.tagID]
-	})
-	for _, cand := range candidates {
-		category := jvNull()
-		if cat := ctx.tagCat[cand.tagID]; cat != "" {
-			category = jvStr(cat)
-		}
-		name := ctx.tagName[cand.tagID]
-		if name == "" {
-			name = cand.tagID
-		}
-		contrast := jvNull()
-		if cand.contrast != nil {
-			contrast = jvFloat(*cand.contrast)
-		}
-		rowsOut.arr = append(rowsOut.arr, jvObj(
-			jvKey("tag_id", jvStr(cand.tagID)),
-			jvKey("name", jvStr(name)),
-			jvKey("category", category),
-			jvKey("cooccurrence", jvInt(cand.cooccur)),
-			jvKey("rate", jvFloat(float64(cand.cooccur)/float64(maxInt64(1, int64(len(baseScenes)))))),
-			jvKey("labeled_n", jvInt(int64(len(withCount(ctx, cand.tagID, labeledBase))))),
-			jvKey("contrast", contrast),
-		))
-	}
-	return jvObj(
-		jvKey("schema_version", jvInt(2)),
-		jvKey("tag_id", jvStr(tagID)),
-		jvKey("items", rowsOut),
-	), nil
-}
-
-func withCount(ctx *curationContext, tagID string, labeledBase map[string]float64) []string {
-	var out []string
-	for sceneID := range labeledBase {
-		if ctx.sceneTags[sceneID][tagID] {
-			out = append(out, sceneID)
-		}
-	}
-	return out
-}
-
-// curationValueStr mirrors str(value) for decoded JSON values in error text.
-func curationValueStr(v jVal) string {
-	switch v.kind {
-	case jStr:
-		return v.s
-	case jNum:
-		return v.num
-	case jBool:
-		if v.b {
-			return "True"
-		}
-		return "False"
-	default:
-		return v.kindName()
-	}
-}
-
-// opGetCurationBatch mirrors backend.py's get_curation_batch branch.
-func opGetCurationBatch(pluginDir string, payload jVal) (jVal, error) {
-	return profiledOperation(pluginDir, payload, "get_curation_batch",
-		func(settings jVal) (jVal, error) { return curationBatchBody(pluginDir, payload, settings) })
-}
-
-func curationBatchBody(pluginDir string, payload, settings jVal) (jVal, error) {
-	args := payload.get("args")
-	mode := pythonStrOrEmpty(args.get("mode"))
-	baseTagID := pythonStrOrEmpty(args.get("base_tag_id"))
-	contextTagID := pythonStrOrEmpty(args.get("context_tag_id"))
-	budget := int(argsInt(args, "budget", curationDefaultBudget))
-	db, err := openAPISidecar(pluginDir, payload, settings)
-	if err != nil {
-		return jvNull(), err
-	}
-	defer db.Close()
-	return createCurationBatch(db, mode, baseTagID, contextTagID, budget)
-}
-
-// opSubmitCurationRatings mirrors backend.py's submit_curation_ratings branch.
-func opSubmitCurationRatings(pluginDir string, payload jVal) (jVal, error) {
-	return profiledOperation(pluginDir, payload, "submit_curation_ratings",
-		func(settings jVal) (jVal, error) { return submitCurationRatingsBody(pluginDir, payload, settings) })
-}
-
-func submitCurationRatingsBody(pluginDir string, payload, settings jVal) (jVal, error) {
-	args := payload.get("args")
-	batchID := pythonStrOrEmpty(args.get("batch_id"))
-	ratings := args.get("ratings")
-	if ratings.kind != jArr {
-		return jvNull(), fmt.Errorf("ratings must be a list")
-	}
-	db, err := openAPISidecar(pluginDir, payload, settings)
-	if err != nil {
-		return jvNull(), err
-	}
-	defer db.Close()
-	return submitCurationRatings(db, batchID, ratings)
-}
-
-// opGetCurationVerdict mirrors backend.py's get_curation_verdict branch.
-func opGetCurationVerdict(pluginDir string, payload jVal) (jVal, error) {
-	return profiledOperation(pluginDir, payload, "get_curation_verdict",
-		func(settings jVal) (jVal, error) { return curationVerdictBody(pluginDir, payload, settings) })
-}
-
-func curationVerdictBody(pluginDir string, payload, settings jVal) (jVal, error) {
-	args := payload.get("args")
-	batchID := pythonStrOrEmpty(args.get("batch_id"))
-	db, err := openAPISidecar(pluginDir, payload, settings)
-	if err != nil {
-		return jvNull(), err
-	}
-	defer db.Close()
-	return curationVerdict(db, batchID)
-}
-
-// opGetTagContextCandidates mirrors backend.py's get_tag_context_candidates branch.
-func opGetTagContextCandidates(pluginDir string, payload jVal) (jVal, error) {
-	return profiledOperation(pluginDir, payload, "get_tag_context_candidates",
-		func(settings jVal) (jVal, error) { return tagContextCandidatesBodyOp(pluginDir, payload, settings) })
-}
-
-func tagContextCandidatesBodyOp(pluginDir string, payload, settings jVal) (jVal, error) {
-	args := payload.get("args")
-	tagID := pythonStrOrEmpty(args.get("tag_id"))
-	minSupport := int(argsInt(args, "min_support", curationDefaultMinSupport))
-	db, err := openAPISidecar(pluginDir, payload, settings)
-	if err != nil {
-		return jvNull(), err
-	}
-	defer db.Close()
-	return tagContextCandidatesBody(db, tagID, minSupport)
-}
-
-// keep the json import used (marshal helpers reference it indirectly).
 
 // ── Pairwise picks (mirror of the curation.py pair section) ────────────────
