+++ curator/curation.py
@@ -1,15 +1,14 @@
-"""Curation batches: hypothesis tests and exploration sampling.
+"""Curation ops: pairwise picks, verdicts, and model impact.
 
-Hypothesis mode draws a stratified 2x2 sample around a (base tag x context
-tag) pair with calibration anchors; explore mode greedily maximizes NEW
-interactive-tag coverage (rarity-weighted, studio-diversity penalty). Ratings
-are stored as ``feedback_type='curation_rating'`` rows carrying the batch and
-cell in payload_json; the verdict is computed from a batch's own ratings
-only, never merged with other labels.
+The pair loop surfaces deterministic pairs of unlabeled scenes across the
+tag, performer, studio, and orthogonal dimensions; each pick answers one
+pair and writes winner/loser labels that feed the model. Pair verdicts
+accumulate per dimension across rounds; curation_impact diffs the two most
+recent model builds.
 
-All selection is deterministic (ORDER BY everywhere, no RNG, explicit
-tie-breaks, half-up rounding only) so the compiled core mirrors every
-function byte-identically; the differential gates enforce that contract.
+Everything is deterministic (ORDER BY, sorted iteration, no RNG, explicit
+tie-breaks) so the compiled core mirrors every function byte-identically;
+the differential gates enforce that contract.
 """
 
 from __future__ import annotations
@@ -18,7 +17,7 @@ import json
 import math
 import sqlite3
 import time
-from collections import Counter, defaultdict, deque
+from collections import Counter, defaultdict
 from dataclasses import dataclass
 from typing import Any, cast
 from uuid import uuid4
@@ -31,7 +30,7 @@ from curator.storage.artifacts import artifact_path, database_path
 
 # StashDB taxonomy categories that describe physical appearance: they cannot
 # participate in interesting tag interactions, so they are excluded from the
-# hypothesis candidate space and from exploration coverage value.
+# interactive tag space.
 EXCLUDED_CATEGORIES = frozenset(
     {
         "Hair Color",
@@ -49,54 +48,7 @@ EXCLUDED_CATEGORIES = frozenset(
     }
 )
 
-MIN_BUDGET, MAX_BUDGET = 1, 40
-DEFAULT_BUDGET = 20
-RATING_MIN, RATING_MAX = 0, 10
-REASON_TYPES = frozenset(
-    {"metadata_wrong", "not_now", "contradicts_hypothesis", "performer_driven"}
-)
-CONFIRM_DELTA = 0.15
-CONFIRM_MIN_N = 10
-ANCHOR_BAND = 200  # scenes each side of the appeal median for calibration anchors
 MAX_ITEM_TAGS = 8
-DEFAULT_MIN_SUPPORT = 20
-CONTRAST_MIN_LABELED = 4  # labeled base scenes needed to rank by outcome contrast
-EXPLORE_ANCHORS = 3
-HYPOTHESIS_ANCHOR_FRACTION = 0.15
-HYPOTHESIS_CONTROL_FRACTION = 0.15
-HYPOTHESIS_CELLS = ("L&T", "L&!T", "!L&T", "!L&!T", "anchor")
-# Context candidates whose library-wide tag rate is at or above this share are
-# ubiquitous (Blowjob, hair colors...): they co-occur with everything and
-# cannot discriminate any relationship.
-CANDIDATE_MAX_LIBRARY_RATE = 0.30
-# Contrast is discounted by label evidence: a candidate's contrast only counts
-# fully once min(with, without) labeled scenes reach this count. Prevents a
-# handful of near-identical labeled scenes from manufacturing phantom
-# relationships (e.g. +0.76 for every co-tag of one rated base tag).
-CONTRAST_EVIDENCE_SCALE = 8.0
-# Hypothesis suggestions skip weak-interaction categories: clothing, mood,
-# location, and technical tags correlate with liked scenes without being
-# meaningful relationships (lesbian + Lingerie), burying real hypotheses like
-# lesbian + Threesome. Acts/Group Makeup/Relations/Finishers stay.
-SUGGESTION_EXCLUDED_CATEGORIES = frozenset(
-    {
-        "Clothing",
-        "Moods",
-        "Locations",
-        "Shot Type",
-        "Surfaces",
-        "Misc",
-        "Accessories",
-    }
-)
-
-
-def _half_up(value: float) -> int:
-    return int(value + 0.5)
-
-
-def _mean(values: list[float]) -> float:
-    return sum(values) / len(values) if values else 0.0
 
 
 @dataclass(frozen=True)
@@ -257,46 +209,6 @@ def curation_context(
     )
 
 
-def _unlabeled_pool(context: CurationContext, scenes: set[str] | frozenset[str]) -> list[str]:
-    return sorted(
-        sid
-        for sid in scenes
-        if sid not in context.labels
-        and sid not in context.blocked_scenes
-        and sid not in context.metadata_wrong
-    )
-
-
-def _round_robin(pool: list[str], quota: int, studio: dict[str, str]) -> list[str]:
-    """One scene per studio before repeats; deterministic order."""
-    by_studio: dict[str, deque[str]] = defaultdict(deque)
-    for sid in pool:
-        by_studio[studio.get(sid, "?")].append(sid)
-    chosen: list[str] = []
-    for name in sorted(by_studio):
-        if len(chosen) >= quota:
-            break
-        chosen.append(by_studio[name].popleft())
-    while len(chosen) < quota:
-        progressed = False
-        for name in sorted(by_studio):
-            if len(chosen) >= quota:
-                break
-            if by_studio[name]:
-                chosen.append(by_studio[name].popleft())
-                progressed = True
-        if not progressed:
-            break
-    return chosen
-
-
-def _anchor_band(pool: list[str], context: CurationContext) -> list[str]:
-    """Appeal middle band: calibration anchors should be average scenes."""
-    ordered = sorted(pool, key=lambda sid: (context.appeal.get(sid, 0.0), sid))
-    mid = len(ordered) // 2
-    return ordered[max(0, mid - ANCHOR_BAND) : mid + ANCHOR_BAND]
-
-
 def _item_tags(context: CurationContext, scene_id: str) -> list[dict[str, str | None]]:
     tags = sorted(
         (t for t in context.scene_tags.get(scene_id, ()) if context.is_interactive(t)),
@@ -321,489 +233,6 @@ def _group_clean_scenes(context: CurationContext, scenes: list[str], tag: str) -
     if context.tag_cat.get(tag) != "Group Makeup":
         return scenes
     return [sid for sid in scenes if len(context.scene_performers.get(sid, ())) < 3]
-
-
-def select_hypothesis(
-    context: CurationContext, base_tag: str, context_tag: str, budget: int
-) -> tuple[list[tuple[str, str, bool]], dict[str, int]]:
-    base = {sid for sid, tags in context.scene_tags.items() if base_tag in tags}
-    ctx = {sid for sid, tags in context.scene_tags.items() if context_tag in tags}
-    all_scenes = context.scene_ids
-    pools = {
-        "L&T": _unlabeled_pool(context, base & ctx),
-        "L&!T": _unlabeled_pool(
-            context, set(_group_clean_scenes(context, list(base - ctx), context_tag))
-        ),
-        "!L&T": _unlabeled_pool(context, ctx - base),
-        "!L&!T": _unlabeled_pool(context, all_scenes - base - ctx),
-    }
-    anchors = max(1, _half_up(budget * HYPOTHESIS_ANCHOR_FRACTION))
-    controls = min(max(1, _half_up(budget * HYPOTHESIS_CONTROL_FRACTION)), budget - anchors)
-    contrast = budget - anchors - controls
-    contrast_l_t = contrast // 2 + contrast % 2
-    contrast_l_nt = contrast // 2
-    items: list[tuple[str, str, bool]] = []
-    for cell, quota in (("L&T", contrast_l_t), ("L&!T", contrast_l_nt), ("!L&T", controls)):
-        items.extend((sid, cell, False) for sid in _round_robin(pools[cell], quota, context.studio))
-    anchor_pool = _anchor_band(pools["!L&!T"], context)
-    items.extend(
-        (sid, "anchor", True) for sid in _round_robin(anchor_pool, anchors, context.studio)
-    )
-    items.sort(key=lambda item: (item[1], item[0]))
-    return items, {cell: len(pool) for cell, pool in pools.items()}
-
-
-def select_explore(
-    context: CurationContext, budget: int
-) -> tuple[list[tuple[str, str, bool]], dict[str, int]]:
-    pool = _unlabeled_pool(context, context.scene_ids)
-    anchors = min(EXPLORE_ANCHORS, max(0, budget // 2))
-    explore_budget = budget - anchors
-    tag_scenes: dict[str, list[str]] = defaultdict(list)
-    for sid in pool:
-        for t in context.scene_tags.get(sid, ()):
-            if context.is_interactive(t):
-                tag_scenes[t].append(sid)
-    value = {
-        sid: sum(
-            context.rarity(t) for t in context.scene_tags.get(sid, ()) if context.is_interactive(t)
-        )
-        for sid in pool
-    }
-    covered: set[str] = set()
-    chosen_set: set[str] = set()
-    chosen_studios: set[str] = set()
-    chosen: list[str] = []
-    while len(chosen) < explore_budget:
-        best, best_value = None, -1.0
-        for sid in sorted(pool):
-            if sid in chosen_set:
-                continue
-            v = value[sid]
-            if context.studio.get(sid, "?") in chosen_studios:
-                v *= 0.5
-            if v > best_value:
-                best, best_value = sid, v
-        if best is None or best_value <= 0:
-            break
-        chosen.append(best)
-        chosen_set.add(best)
-        chosen_studios.add(context.studio.get(best, "?"))
-        for t in context.scene_tags.get(best, ()):
-            if context.is_interactive(t) and t not in covered:
-                covered.add(t)
-                for other in tag_scenes[t]:
-                    if other in value:
-                        value[other] -= context.rarity(t)
-    anchor_pool = _anchor_band([sid for sid in pool if sid not in chosen_set], context)
-    items = [(sid, "explore", False) for sid in chosen]
-    items.extend(
-        (sid, "anchor", True) for sid in _round_robin(anchor_pool, anchors, context.studio)
-    )
-    items.sort(key=lambda item: (item[1], item[0]))
-    return items, {"candidates": len(pool), "interactive_tags": len(context.interactive)}
-
-
-def create_batch(
-    connection: sqlite3.Connection,
-    mode: str,
-    base_tag_id: str | None,
-    context_tag_id: str | None,
-    budget: int,
-    config: CuratorConfig = DEFAULT_CONFIG,
-) -> dict[str, object]:
-    if mode not in ("hypothesis", "explore"):
-        raise ValueError("mode must be 'hypothesis' or 'explore'")
-    if not MIN_BUDGET <= budget <= MAX_BUDGET:
-        raise ValueError(f"budget must be from {MIN_BUDGET} to {MAX_BUDGET}")
-    if mode == "hypothesis":
-        if not base_tag_id or not context_tag_id:
-            raise ValueError("hypothesis mode requires base_tag_id and context_tag_id")
-        if base_tag_id == context_tag_id:
-            raise ValueError("base_tag_id and context_tag_id must differ")
-    for tag_id in (base_tag_id, context_tag_id):
-        if (
-            tag_id is not None
-            and connection.execute("SELECT 1 FROM source_tag WHERE tag_id=?", (tag_id,)).fetchone()
-            is None
-        ):
-            raise ValueError(f"unknown tag: {tag_id}")
-    context = curation_context(connection, config)
-    if mode == "hypothesis":
-        assert base_tag_id is not None and context_tag_id is not None
-        items, pool = select_hypothesis(context, base_tag_id, context_tag_id, budget)
-        anchors = sum(1 for _, _, anchor in items if anchor)
-        policy = (
-            f"stratified 2x2, studio round-robin, unlabeled only, {anchors} calibration anchors"
-        )
-    else:
-        items, pool = select_explore(context, budget)
-        policy = "max-coverage interactive tags, rarity-weighted, studio penalty, unlabeled only"
-    batch_id = str(uuid4())
-    now_ms = time.time_ns() // 1_000_000
-    with transaction(connection):
-        connection.execute(
-            """
-            INSERT INTO curation_batch(
-                batch_id, mode, base_tag_id, context_tag_id, budget, status,
-                created_at_ms, payload_json
-            ) VALUES (?, ?, ?, ?, ?, 'open', ?, ?)
-            """,
-            (
-                batch_id,
-                mode,
-                base_tag_id,
-                context_tag_id,
-                budget,
-                now_ms,
-                json.dumps({"policy": policy, "pool": pool}, sort_keys=True, separators=(",", ":")),
-            ),
-        )
-        connection.executemany(
-            """
-            INSERT INTO curation_batch_item(batch_id, scene_id, cell, anchor)
-            VALUES (?, ?, ?, ?)
-            """,
-            [(batch_id, sid, cell, 1 if anchor else 0) for sid, cell, anchor in items],
-        )
-    return {
-        "schema_version": 2,
-        "batch_id": batch_id,
-        "mode": mode,
-        "base_tag_id": base_tag_id,
-        "context_tag_id": context_tag_id,
-        "budget": budget,
-        "items": [
-            {
-                "scene_id": sid,
-                "cell": cell,
-                "anchor": anchor,
-                "title": context.scene_title.get(sid),
-                "studio": context.studio.get(sid),
-                "date": context.scene_date.get(sid),
-                "details": context.scene_details.get(sid),
-                "tags": _item_tags(context, sid),
-            }
-            for sid, cell, anchor in items
-        ],
-        "pool": pool,
-        "policy": policy,
-    }
-
-
-def _batch_items(connection: sqlite3.Connection, batch_id: str) -> dict[str, tuple[str, bool]]:
-    return {
-        str(row["scene_id"]): (str(row["cell"]), bool(row["rated"]))
-        for row in connection.execute(
-            "SELECT scene_id, cell, rated FROM curation_batch_item WHERE batch_id=?",
-            (batch_id,),
-        )
-    }
-
-
-def _insert_reason_feedback(
-    connection: sqlite3.Connection, now_ms: int, scene_id: str, reason: str
-) -> None:
-    connection.execute(
-        """
-        INSERT INTO feedback(
-            feedback_id, scene_id, feedback_type, value, occurred_at_ms,
-            reversed_by_id, impression_id, payload_json
-        ) VALUES (?, ?, ?, NULL, ?, NULL, NULL, '{}')
-        """,
-        (f"{now_ms}-{uuid4().hex}", scene_id, reason, now_ms),
-    )
-
-
-def submit_ratings(
-    connection: sqlite3.Connection, batch_id: str, ratings: list[dict[str, Any]]
-) -> dict[str, object]:
-    if not batch_id:
-        raise ValueError("batch_id is required")
-    batch = connection.execute(
-        "SELECT status FROM curation_batch WHERE batch_id=?", (batch_id,)
-    ).fetchone()
-    if batch is None:
-        raise ValueError(f"unknown batch: {batch_id}")
-    if str(batch["status"]) != "open":
-        raise ValueError(f"batch is not open: {batch['status']}")
-    items = _batch_items(connection, batch_id)
-    if not items:
-        raise ValueError(f"unknown batch: {batch_id}")
-    seen: set[str] = set()
-    normalized: list[tuple[str, int, str | None]] = []
-    for entry in ratings:
-        if not isinstance(entry, dict):
-            raise ValueError("each rating must be an object with scene_id, value, and reason")
-        scene_id = str(entry.get("scene_id") or "")
-        if scene_id not in items:
-            raise ValueError(f"scene is not in this batch: {scene_id}")
-        if scene_id in seen:
-            raise ValueError(f"duplicate rating for scene: {scene_id}")
-        if items[scene_id][1]:
-            raise ValueError(f"scene already rated in this batch: {scene_id}")
-        raw_value = entry.get("value")
-        if isinstance(raw_value, bool) or not isinstance(raw_value, int):
-            raise ValueError("rating value must be an integer")
-        if not RATING_MIN <= raw_value <= RATING_MAX:
-            raise ValueError(f"rating value must be from {RATING_MIN} to {RATING_MAX}")
-        reason = entry.get("reason")
-        if reason is not None and reason not in REASON_TYPES:
-            raise ValueError(f"unknown rating reason: {reason}")
-        seen.add(scene_id)
-        normalized.append((scene_id, raw_value, str(reason) if reason else None))
-    if not normalized:
-        raise ValueError("ratings must not be empty")
-    now_ms = time.time_ns() // 1_000_000
-    with transaction(connection):
-        for scene_id, value, reason in normalized:
-            cell = items[scene_id][0]
-            payload = {"batch_id": batch_id, "cell": cell}
-            connection.execute(
-                """
-                INSERT INTO feedback(
-                    feedback_id, scene_id, feedback_type, value, occurred_at_ms,
-                    reversed_by_id, impression_id, payload_json
-                ) VALUES (?, ?, 'curation_rating', ?, ?, NULL, NULL, ?)
-                """,
-                (
-                    f"{now_ms}-{uuid4().hex}",
-                    scene_id,
-                    str(value),
-                    now_ms,
-                    json.dumps(payload, sort_keys=True, separators=(",", ":")),
-                ),
-            )
-            if reason is not None:
-                _insert_reason_feedback(connection, now_ms, scene_id, reason)
-            connection.execute(
-                "UPDATE curation_batch_item SET rated=1 WHERE batch_id=? AND scene_id=?",
-                (batch_id, scene_id),
-            )
-        remaining = int(
-            connection.execute(
-                "SELECT count(*) FROM curation_batch_item WHERE batch_id=? AND rated=0",
-                (batch_id,),
-            ).fetchone()[0]
-        )
-        status = "open" if remaining else "rated"
-        if status == "rated":
-            connection.execute(
-                "UPDATE curation_batch SET status='rated' WHERE batch_id=?", (batch_id,)
-            )
-    return {
-        "schema_version": 2,
-        "accepted": len(normalized),
-        "batch_status": status,
-    }
-
-
-def _batch_ratings(connection: sqlite3.Connection, batch_id: str) -> dict[str, float]:
-    """This batch's own curation ratings: scene_id -> outcome.
-
-    Ratings marked with a reason are excluded from the tag-level verdict:
-    metadata_wrong / contradicts_hypothesis scenes are not valid instances of
-    the cell they were sorted into, and performer_driven ratings would spill
-    performer preference into every tag's mean. All of them still count as
-    model labels (the model wants performer outcomes for performer affinity).
-    """
-    excluded = {
-        str(row["scene_id"])
-        for row in connection.execute(
-            """SELECT DISTINCT scene_id FROM feedback
-               WHERE feedback_type IN
-                     ('metadata_wrong', 'contradicts_hypothesis', 'performer_driven')
-                 AND reversed_by_id IS NULL"""
-        )
-    }
-    outcomes: dict[str, float] = {}
-    for row in connection.execute(
-        """SELECT scene_id, value, payload_json FROM feedback
-           WHERE feedback_type='curation_rating' AND reversed_by_id IS NULL"""
-    ):
-        payload = json.loads(str(row["payload_json"]) or "{}")
-        if payload.get("batch_id") != batch_id:
-            continue
-        if str(row["scene_id"]) in excluded:
-            continue
-        try:
-            rating = int(row["value"])
-        except (TypeError, ValueError):
-            continue
-        if not RATING_MIN <= rating <= RATING_MAX:
-            continue
-        outcomes[str(row["scene_id"])] = (rating - 5) / 5
-    return outcomes
-
-
-def _cell_stats(
-    outcomes: dict[str, float], items: dict[str, tuple[str, bool]]
-) -> list[dict[str, object]]:
-    rows: list[dict[str, object]] = []
-    for cell in HYPOTHESIS_CELLS:
-        values = [outcomes[sid] for sid, (c, _) in items.items() if c == cell and sid in outcomes]
-        rows.append(
-            {
-                "cell": cell,
-                "n": len(values),
-                "mean_outcome": _mean(values) if values else None,
-            }
-        )
-    return rows
-
-
-def verdict(connection: sqlite3.Connection, batch_id: str) -> dict[str, object]:
-    if not batch_id:
-        raise ValueError("batch_id is required")
-    batch = connection.execute(
-        "SELECT mode, base_tag_id, context_tag_id FROM curation_batch WHERE batch_id=?",
-        (batch_id,),
-    ).fetchone()
-    if batch is None:
-        raise ValueError(f"unknown batch: {batch_id}")
-    mode = str(batch["mode"])
-    items = _batch_items(connection, batch_id)
-    outcomes = _batch_ratings(connection, batch_id)
-    if mode == "hypothesis":
-        cells = _cell_stats(outcomes, items)
-        by_cell = {str(c["cell"]): c for c in cells}
-        contrast: dict[str, object] = {}
-        l_t, l_nt = by_cell["L&T"], by_cell["L&!T"]
-        if l_t["n"] and l_nt["n"]:
-            mean_lt = cast(float, l_t["mean_outcome"])
-            mean_lnt = cast(float, l_nt["mean_outcome"])
-            delta = mean_lt - mean_lnt
-            n_total = int(cast(int, l_t["n"])) + int(cast(int, l_nt["n"]))
-            contrast = {
-                "delta": delta,
-                "n_total": n_total,
-                "confirmed": abs(delta) >= CONFIRM_DELTA and n_total >= CONFIRM_MIN_N,
-            }
-        suggested: dict[str, object] | None = None
-        if l_t["n"]:
-            mean_lt = cast(float, l_t["mean_outcome"])
-            value = max(-1.0, min(1.0, _half_up(mean_lt * 2) / 2))
-            suggested = {
-                "base_tag_id": str(batch["base_tag_id"] or ""),
-                "context_tag_id": str(batch["context_tag_id"] or ""),
-                "value": value,
-            }
-        return {
-            "schema_version": 2,
-            "batch_id": batch_id,
-            "mode": mode,
-            "cells": cells,
-            "contrast": contrast,
-            "suggested_rule": suggested,
-        }
-    # explore mode: top/bottom tags by mean outcome across the batch's ratings.
-    rated_scenes = set(outcomes)
-    tag_rows: dict[str, list[float]] = defaultdict(list)
-    if rated_scenes:
-        placeholders = ",".join("?" * len(rated_scenes))
-        for row in connection.execute(
-            f"""SELECT scene_id, tag_id FROM scene_tag
-                WHERE scene_id IN ({placeholders}) ORDER BY scene_id, tag_id""",
-            tuple(rated_scenes),
-        ):
-            tag_rows[str(row["tag_id"])].append(outcomes[str(row["scene_id"])])
-    tag_names: dict[str, str] = {}
-    tag_cats: dict[str, str] = {}
-    context = curation_context(connection)
-    for tag_id in tag_rows:
-        tag_names[tag_id] = context.tag_name.get(tag_id, tag_id)
-        tag_cats[tag_id] = context.tag_cat.get(tag_id, "") or ""
-    entries = [
-        {
-            "tag_id": tag_id,
-            "name": tag_names[tag_id],
-            "category": tag_cats[tag_id] or None,
-            "n": len(values),
-            "mean_outcome": _mean(values),
-        }
-        for tag_id, values in tag_rows.items()
-        if len(values) >= 2
-    ]
-    entries.sort(key=lambda item: (-cast(float, item["mean_outcome"]), str(item["tag_id"])))
-    values = list(outcomes.values())
-    return {
-        "schema_version": 2,
-        "batch_id": batch_id,
-        "mode": mode,
-        "summary": {"n": len(values), "mean_outcome": _mean(values) if values else None},
-        "top_tags": entries[:10],
-        "bottom_tags": entries[-10:][::-1],
-    }
-
-
-def tag_context_candidates(
-    connection: sqlite3.Connection,
-    tag_id: str,
-    min_support: int = DEFAULT_MIN_SUPPORT,
-    config: CuratorConfig = DEFAULT_CONFIG,
-) -> dict[str, object]:
-    if not tag_id:
-        raise ValueError("tag_id is required")
-    if connection.execute("SELECT 1 FROM source_tag WHERE tag_id=?", (tag_id,)).fetchone() is None:
-        raise ValueError(f"unknown tag: {tag_id}")
-    if min_support < 1:
-        raise ValueError("min_support must be at least 1")
-    context = curation_context(connection, config)
-    labels = PreferenceModelBuilder(connection, config)._scene_labels()
-    base_scenes = {sid for sid, tags in context.scene_tags.items() if tag_id in tags}
-    labeled_base = {sid: labels[sid].outcome for sid in base_scenes if sid in labels}
-    cooc: Counter[str] = Counter()
-    for sid in base_scenes:
-        for t in context.scene_tags.get(sid, ()):
-            if t != tag_id and context.is_interactive(t):
-                cooc[t] += 1
-    rows: list[dict[str, object]] = []
-    total_scenes = max(1, len(context.scene_ids))
-    for t, n in cooc.items():
-        if n < min_support:
-            continue
-        if context.counts.get(t, 0) / total_scenes >= CANDIDATE_MAX_LIBRARY_RATE:
-            continue  # ubiquitous tags cannot discriminate any relationship
-        if context.tag_name.get(t, t).startswith("["):
-            continue  # sync-artifact tags ([Timestamp: Synced]...) are junk
-        if context.tag_cat.get(t, "") in SUGGESTION_EXCLUDED_CATEGORIES:
-            continue  # weak-interaction categories are not hypotheses
-        with_t = [labeled_base[sid] for sid in labeled_base if t in context.scene_tags.get(sid, ())]
-        without_t = [
-            labeled_base[sid]
-            for sid in _group_clean_scenes(
-                context,
-                [sid for sid in labeled_base if t not in context.scene_tags.get(sid, ())],
-                t,
-            )
-        ]
-        contrast: float | None = None
-        if len(labeled_base) >= CONTRAST_MIN_LABELED and with_t and without_t:
-            raw = _mean(with_t) - _mean(without_t)
-            evidence = min(len(with_t), len(without_t))
-            contrast = raw * min(1.0, evidence / CONTRAST_EVIDENCE_SCALE)
-        rows.append(
-            {
-                "tag_id": t,
-                "name": context.tag_name.get(t, t),
-                "category": context.tag_cat.get(t, "") or None,
-                "cooccurrence": n,
-                "rate": n / max(1, len(base_scenes)),
-                "labeled_n": len(with_t),
-                "contrast": contrast,
-            }
-        )
-    rows.sort(
-        key=lambda item: (
-            # Measured POSITIVE contrasts first (rescue contexts); null and
-            # negative contrasts sink below them.
-            0 if item["contrast"] is not None and cast(float, item["contrast"]) > 0 else 1,
-            -(cast(float, item["contrast"]) if item["contrast"] is not None else 0.0),
-            -int(cast(int, item["cooccurrence"])),
-            str(item["name"]),
-        )
-    )
-    return {"schema_version": 2, "tag_id": tag_id, "items": rows}
 
 
 # ── Pairwise picks ───────────────────────────────────────────────────────────
